requirements.txt
pyyaml
pytest
yamllint
